+++ src/models/Occurrence.php
@@ -10,6 +10,8 @@
 
 namespace unionco\calendarize\models;
 
+use Craft;
+use craft\helpers\FileHelper;
 use craft\base\Element;
 use DateTime;
 use ReflectionClass;