+++ src/variables/CalendarizeVariable.php
@@ -14,6 +14,7 @@ use Craft;
 use craft\elements\Entry;
 use craft\elements\MatrixBlock;
 use unionco\calendarize\Calendarize;
+use unionco\calendarize\models\CalendarizeModel;
 
 /**
  * @author    Franco Valdes